feat(cpkernel): get node effusivity
Adds method to get the node effusivity. This method is only implemented for
the debris treatment to avoid bloat in clean-ice simulations.

The method can be used to solve for the contact temperature between
debris and snow/ice, and hence provide BCs for solving the heat equation
within the debris layers.